requirements.txt
# PixelStitch 项目依赖

# 深度学习框架
torch>=1.8.0
torchvision>=0.9.0

# 数值计算和科学计算
numpy>=1.19.0
scipy>=1.5.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0

# 可视化和日志
matplotlib>=3.3.0
tensorboard>=2.4.0

# 进度条
tqdm>=4.60.0

